+++ src/Views/NullView.php
@@ -11,43 +11,17 @@ declare(strict_types=1);
 
 namespace App\Views;
 
-use App\Models\HomeModel;
-use App\Templates\HtmlTemplate;
-use Linna\Authentication\Authentication;
 use Linna\Mvc\View;
 
 /**
- * Home Page View.
+ * Null View.
  */
-class HomeView extends View
+class NullView extends View
 {
-    /**
-     * Constructor.
-     *
-     * @param HomeModel      $model
-     * @param Authentication $login
-     * @param HtmlTemplate   $htmlTemplate
-     */
-    public function __construct(HomeModel $model, Authentication $login, HtmlTemplate $htmlTemplate)
-    {
-        parent::__construct($model);
-
-        //merge data passed from model with login information
-        $this->data = array_merge($this->data, ['login' => $login->islogged(), 'userName' => $login->getLoginData()['user_name']]);
-
-        //store html template
-        $this->template = $htmlTemplate;
-    }
-
     /**
      * Index.
      */
-    public function index()
+    public function index(): void
     {
-        //load home html
-        $this->template->loadHtml('Home');
-
-        //set page title
-        $this->template->title = 'App/Home';
     }
 }